+++ src/main/resources/db/migration/V7_1__create_shows_table.sql
@@ -1,5 +1,5 @@
 CREATE TABLE `shows` (
-  `id` int(11) NOT NULL AUTO_INCREMENT,
+  `id` BIGINT NOT NULL AUTO_INCREMENT,
   `slug` varchar(60) COLLATE utf8mb4_unicode_ci NOT NULL UNIQUE,
   `title` varchar(255) COLLATE utf8mb4_unicode_ci NOT NULL,
   `description` longtext,
@@ -8,9 +8,9 @@ CREATE TABLE `shows` (
   `price` decimal(10,2) DEFAULT NULL,
   `created_at` datetime(6) NOT NULL,
   `updated_at` datetime(6) DEFAULT NULL,
-  `location_id` int(11) DEFAULT NULL,
+  `location_id` BIGINT DEFAULT NULL,
   PRIMARY KEY(id)
-) ENGINE=InnoDB DEFAULT CHARSET= utf8mb4 COLLATE=utf8mb4_unicode_ci;
+) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_ci;
 
 --
 -- Index pour la table `shows`
